Committing all changes.
Key shift was using UserLoginSerializer instead of UserSerializer. This
is because the former has the definition of token in it. I added all the
values I need to come through (avatar, specialist_id, etc.) to this file
and kept the token definition. So hopefully this will work now.

At this point, the one functionality I want that I don't have is
automatically assigning a new user to a specialist based on their
wing. There is some code in another branch for this which I am going to try
to find.